fixed pageNavigation WIP for ComponentEvaluation feature
