--- RKCIUIAutomation/Page/Navigation/INavMenu.cs
@@ -1,14 +1,0 @@
-﻿using RKCIUIAutomation.Page.PageObjects.Project;
-
-namespace RKCIUIAutomation.Page.Navigation
-{
-    interface INavMenu_Project
-    {
-        //Project NavMenu Interfaces
-
-        MyDetails NavigateToMyDetails();
-        QmsDocuments NavigateToQmsDocuments();
-        ProjectDetails NavigateToProjectDetails();
-    }
-
-}
