+++ RKCIUIAutomation/RKCIUIAutomation.csproj
@@ -138,12 +138,10 @@
     <Compile Include="Config\ProjectProperties.cs" />
     <Compile Include="Config\Properties.cs" />
     <Compile Include="Page\Action.cs" />
-    <Compile Include="Page\PageObjects\LandingPage.cs" />
-    <Compile Include="Page\Navigation\INavMenu.cs" />
     <Compile Include="Page\Navigation\NavMenu.cs" />
+    <Compile Include="Page\PageObjects\LandingPage.cs" />
     <Compile Include="Page\PageBase.cs" />
     <Compile Include="Page\PageObjects\LoginPage.cs" />
-    <Compile Include="Page\Navigation\NavEnums.cs" />
     <Compile Include="Page\PageHelper.cs" />
     <Compile Include="Page\PageObjects\Project\MyDetails.cs" />
     <Compile Include="Page\PageObjects\Project\Project.cs" />
